+++ tests/src/cbor/cbor_view_tests.cpp
@@ -217,21 +217,21 @@ TEST_CASE("test_dump_to_string")
 
     std::string s1;
     json_options options1;
-    options1.bignum_format(bignum_chars_format::integer);
+    options1.big_integer_format(big_integer_chars_format::number);
     bv.dump(s1,options1);
     CHECK("[-18446744073709551617]" == s1);
     //std::cout << s1 << std::endl;
 
     std::string s2;
     json_options options2;
-    options2.bignum_format(bignum_chars_format::base10);
+    options2.big_integer_format(big_integer_chars_format::base10);
     bv.dump(s2,options2);
     CHECK("[\"-18446744073709551617\"]" == s2);
     //std::cout << s2 << std::endl;
 
     std::string s3;
     json_options options3;
-    options3.bignum_format(bignum_chars_format::base64url);
+    options3.big_integer_format(big_integer_chars_format::base64url);
     bv.dump(s3,options3);
     CHECK("[\"~AQAAAAAAAAAA\"]" == s3);
     //std::cout << s3 << std::endl;
@@ -259,21 +259,21 @@ TEST_CASE("test_dump_to_stream")
 
     std::ostringstream os1;
     json_options options1;
-    options1.bignum_format(bignum_chars_format::integer);
+    options1.big_integer_format(big_integer_chars_format::number);
     bv.dump(os1,options1);
     CHECK("[-18446744073709551617]" == os1.str());
     //std::cout << os1.str() << std::endl;
 
     std::ostringstream os2;
     json_options options2;
-    options2.bignum_format(bignum_chars_format::base10);
+    options2.big_integer_format(big_integer_chars_format::base10);
     bv.dump(os2,options2);
     CHECK("[\"-18446744073709551617\"]" == os2.str());
     //std::cout << os2.str() << std::endl;
 
     std::ostringstream os3;
     json_options options3;
-    options3.bignum_format(bignum_chars_format::base64url);
+    options3.big_integer_format(big_integer_chars_format::base64url);
     bv.dump(os3,options3);
     CHECK("[\"~AQAAAAAAAAAA\"]" == os3.str());
     //std::cout << os3.str() << std::endl;